--- a/setup.py
+++ b/setup.py
@@ -261,6 +261,7 @@ def run_setup(with_binary):
             "numba",  # just-in-time compiler
             "psutil",  # for getting user RAM
             "pyyaml",  # for printing profiler dictionary
+            "tuna",  # to generate visual/interactive performance profiles
         ],
         extras_require={
             "dev": [