<!-- Please be sure to check out our contributing guidelines,
https://github.com/radis/radis/blob/develop/CONTRIBUTING.md . -->

### Description
<!-- Provide a general description of what your pull request does. -->

use [tuna](https://github.com/nschloe/tuna) to generate visual/interactive performance profiles; and the update the console profiler:


Syntax : 

```
s = calc_spectrum(...)
s.generate_perf_profile()
```


Typical output (opens a local web page) : 

![image](https://user-images.githubusercontent.com/16088743/128018032-6049be72-1881-46ac-9d7c-1ed89f9c4f42.png)

Console profiler:

```
s = calc_spectrum(...)
s.print_perf_profile()
```

Typical output : 

```
        spectrum_calculation      0.844s ████████████████
            check_line_databank              0.001s
            check_non_eq_param               0.070s █
            fetch_energy_5                   0.027s
            calc_weight_trans                0.016s
            reinitialize                     0.003s
                copy_database                    0.000s
                memory_usage_warning             0.003s
                reset_population                 0.000s
            calc_noneq_population            0.066s █
                part_function                    0.056s █
                population                       0.010s
            scaled_non_eq_linestrength       0.004s
                map_part_func                    0.001s
                corrected_population_se          0.003s
            calc_emission_integral           0.008s
            applied_linestrength_cutoff      0.003s
            calc_lineshift                   0.001s
            calc_hwhm                        0.008s
            generate_wavenumber_arrays       0.002s
            calc_line_broadening             0.668s ████████████
                precompute_DLM_lineshapes        0.020s
                DLM_Initialized_vectors          0.000s
                DLM_closest_matching_line        0.001s
                DLM_Distribute_lines             0.002s
                DLM_convolve                     0.304s █████
            calc_other_spectral_quan         0.005s
            generate_spectrum_obj            0.000s
```

### Implementation

- [x] Proof of concept
- [x] Test
- [x] update "profiler" format consistently : @anandxkumar see below 
- [x] mention in docs/performances


@anandxkumar this expects the `conditions["profiler"]` dict to be of the format:  
```python
prof = {'scaled_eq_linestrength': 0.02393484115600586, 'calc_lineshift': 0.002992868423461914, 'calc_hwhm': 0.04986977577209473, 'precompute_DLM_lineshapes': 0.06680488586425781, 'calc_line_broadening': {'value': 1.423218011856079, 'DLM_Initialized_vectors': 0.0, 'DLM_closest_matching_line': 0.048909902572631836, 'DLM_Distribute_lines': 0.052854061126708984, 'DLM_convolve': 1.2426400184631348}, 'calc_other_spectral_quan': 0.03287649154663086, 'spectrum_calc_before_obj': 1.560823678970337, 'value': 1.560823678970337}
```
i.e. (more human-readable with `print(yaml.dump(prof, allow_unicode=True, default_flow_style=False))`
```
calc_hwhm: 0.04986977577209473
calc_line_broadening:
  DLM_Distribute_lines: 0.052854061126708984
  DLM_Initialized_vectors: 0.0
  DLM_closest_matching_line: 0.048909902572631836
  DLM_convolve: 1.2426400184631348
  value: 1.423218011856079              👈  notice the "value"
calc_lineshift: 0.002992868423461914
calc_other_spectral_quan: 0.03287649154663086
precompute_DLM_lineshapes: 0.06680488586425781
scaled_eq_linestrength: 0.02393484115600586
spectrum_calc_before_obj: 1.560823678970337
value: 1.560823678970337          👈  notice the "value"
```

This PR generates the corresponding Python Tree. 

I suggest that you use this PR to also close #332 ; 

- [x] maybe also add a Spectrum method using the `yaml.dump` function above to nicely print the profiler dictionary, adding % etc ?