@anandxkumar 2021-08-03
@erwanp  Indentation Error:
```

sf = SpectrumFactory(wavenum_min=2900,
                     wavenum_max=4200,
                     molecule='OH',
                     broadening_max_width=10,  # cm-1
                     medium='vacuum',
                     verbose=3,    # more for more details
                     pressure = 10,
                     wstep = 0.01,
                     )
sf.warnings={"AccuracyError":"ignore",
             "MissingSelfBroadeningWarning":"ignore"}
sf.fetch_databank('hitemp')
s1 = sf.eq_spectrum(Tgas=300, path_length=1)
s1.print_perf_profile()
```

Output:
```
        spectrum_calculation  0.646s ████████████████
        check_line_databank         0.000s 
            reinitialize                0.008s 
            copy_database               0.001s 
            memory_usage_warning        0.001s 
            reset_population            0.000s 
        scaled_eq_linestrength      0.003s 
        calc_lineshift              0.001s 
        calc_hwhm                   0.004s 
        generate_wavenumber_arrays  0.003s 
            calc_line_broadening        0.618s ███████████████
            precompute_DLM_lineshapes   0.024s 
            DLM_Initialized_vectors     0.000s 
            DLM_closest_matching_line   0.001s 
            DLM_Distribute_lines        0.001s 
            DLM_convolve                0.591s ██████████████
        calc_other_spectral_quan    0.006s 
        generate_spectrum_obj       0.001s 
 ```
 
`reinitialize` and `calc_line_broadening` should be at the previous level indentation.